Rename evals/eval.py module to evals/basis_eval.py
Change the code, docstrings, notes and file names to reflect this renaming.
