+++ gbasis/evals/density.py
@@ -1,5 +1,5 @@
 """Density Evaluation."""
-from gbasis.evals.eval import evaluate_basis
+from gbasis.evals.basis_eval import evaluate_basis
 from gbasis.evals.eval_deriv import evaluate_deriv_basis
 import numpy as np
 from scipy.special import comb
