+++ tests/test_basis_eval.py
@@ -1,7 +1,7 @@
-"""Test gbasis.evals.eval."""
+"""Test gbasis.evals.basis_eval."""
 from gbasis.contractions import GeneralizedContractionShell
 from gbasis.evals._deriv import _eval_deriv_contractions
-from gbasis.evals.eval import Eval, evaluate_basis
+from gbasis.evals.basis_eval import Eval, evaluate_basis
 from gbasis.parsers import make_contractions, parse_nwchem
 import numpy as np
 import pytest
@@ -10,7 +10,7 @@ from utils import find_datafile, HortonContractions
 
 
 def test_evaluate_construct_array_contraction():
-    """Test gbasis.evals.eval.Eval.construct_array_contraction."""
+    """Test gbasis.evals.basis_eval.Eval.construct_array_contraction."""
     test = GeneralizedContractionShell(
         1, np.array([0.5, 1, 1.5]), np.array([1.0, 2.0]), np.array([0.1, 0.01])
     )
@@ -54,7 +54,7 @@ def test_evaluate_construct_array_contraction():
 
 
 def test_evaluate_basis_cartesian():
-    """Test gbasis.evals.eval.evaluate_basis_cartesian."""
+    """Test gbasis.evals.basis_eval.evaluate_basis with cartesian coordinate type."""
     basis_dict = parse_nwchem(find_datafile("data_sto6g.nwchem"))
     basis = make_contractions(basis_dict, ["H"], np.array([[0, 0, 0]]))
     evaluate_obj = Eval(basis)
@@ -65,7 +65,7 @@ def test_evaluate_basis_cartesian():
 
 
 def test_evaluate_basis_spherical():
-    """Test gbasis.evals.eval.evaluate_basis_spherical."""
+    """Test gbasis.evals.basis_eval.evaluate_basis with spherical coordinate type."""
     basis_dict = parse_nwchem(find_datafile("data_sto6g.nwchem"))
 
     # cartesian and spherical are the same for s orbital
@@ -92,7 +92,7 @@ def test_evaluate_basis_spherical():
 
 
 def test_evaluate_basis_mix():
-    """Test gbasis.evals.eval.evaluate_basis_mix."""
+    """Test gbasis.evals.basis_eval.evaluate_basis with mix coordinate type."""
     basis_dict = parse_nwchem(find_datafile("data_sto6g.nwchem"))
 
     # cartesian and spherical are the same for s orbital
@@ -118,7 +118,7 @@ def test_evaluate_basis_mix():
 
 
 def test_evaluate_basis_lincomb():
-    """Test gbasis.evals.eval.evaluate_basis_lincomb."""
+    """Test gbasis.evals.basis_eval.evaluate_basis with lincomb."""
     basis_dict = parse_nwchem(find_datafile("data_sto6g.nwchem"))
     basis = make_contractions(basis_dict, ["Kr"], np.array([[0, 0, 0]]))
     evaluate_obj = Eval(basis)
@@ -130,7 +130,7 @@ def test_evaluate_basis_lincomb():
 
 
 def test_evaluate_basis_horton():
-    """Test gbasis.evals.eval.evaluate_basis against horton results."""
+    """Test gbasis.evals.basis_eval.evaluate_basis against horton results."""
     basis_dict = parse_nwchem(find_datafile("data_anorcc.nwchem"))
     points = np.array([[0, 0, 0], [0.8, 0, 0]])
     basis = make_contractions(basis_dict, ["H", "He"], points)
@@ -148,7 +148,7 @@ def test_evaluate_basis_horton():
 
 
 def test_evaluate_basis_pyscf():
-    """Test gbasis.evals.eval.evaluate_basis against pyscf results."""
+    """Test gbasis.evals.basis_eval.evaluate_basis against pyscf results."""
     pytest.importorskip("pyscf")
 
     from pyscf import gto
@@ -209,7 +209,7 @@ def test_evaluate_basis_pyscf():
 
 @pytest.mark.xfail
 def test_evaluate_basis_pyscf_cart_norm():
-    """Test gbasis.evals.eval.evaluate_basis against pyscf results.
+    """Test gbasis.evals.basis_eval.evaluate_basis against pyscf results.
 
     These cases fail because pyscf seems to have a different normalization constant for the d and f
     orbitals.
